refactor!: remove named export 'parse()'
This is an internal method and never meant to be called by users.
Furthermore, the default export `Psd` class also has a static `parse()`
method, which could be easily confused with the method.
Thus, we're removing it from the list of exports.

Also edit Node.js example script, which was previously using this
`parse()` to print information to the console (indicating successful
parse). Instead, stringify and print the `Psd` object directly.