chore: update rimraf, update NPM 'clear' script
Update rimraf to 4.4.0. Now that rimraf requires `--glob` to explicitly
expand wildcards (`*`), update the 'clear' script accordingly.
This isn't really needed for on macOS and Linux, where command shells
expand globs anyway. We have been unknowingly relying on this behavior.
But it's good to provide the correct solution when we can.